+++ src/server/app-rest-preset.ts
@@ -53,6 +53,8 @@ export default class RESTPreset {
                             if (ducky.validate(obj, StateSchemaPartial))
                                 StateUtil.copy(state, obj)
                         }
+                        if (state.renderer   !== undefined) n++
+                        if (state.streams    !== undefined) n++
                         if (state.canvas     !== undefined) n++
                         if (state.monitor    !== undefined) n++
                         if (state.decal      !== undefined) n++
@@ -65,7 +67,6 @@ export default class RESTPreset {
                         if (state.CAM2       !== undefined) n++
                         if (state.CAM3       !== undefined) n++
                         if (state.CAM4       !== undefined) n++
-                        if (state.renderer   !== undefined) n++
                         presets.push(n)
                     }
                     return h.response(presets).code(200)